Apply changes in comments made by Dr. Coleman.
Also, figured out why the csv file was not being recognized.
Depending on where we call pytest from in the in terminal (i.e. what folder you are currently in, in the structure), pytest will not be able to see the csv file unless it is in that folder or the path is defined from there.
Currently, I moved it into the main clashboard folder but we can change that if we decide a better location. There is likely a workaround to making that not a problem from running pytest anywhere in the directory structure but I did not know how to do that.